`Refactored AdminController and related views and routes to remove unused code and improve structure.`

+++ routes/web.php
@@ -7,7 +7,6 @@ use App\Http\Controllers\FavoriteController;
 use App\Http\Controllers\DashboardController;
 use App\Http\Controllers\BookController;
 use App\Http\Controllers\AdminController;
-use App\Http\Controllers\Admin\GenreController;
 use Illuminate\Support\Facades\Route;
 
 Route::get('/', function () {
@@ -50,32 +49,27 @@ Route::middleware('auth')->group(function () {
     Route::patch('/profile', [ProfileController::class, 'update'])->name('profile.update');
     Route::delete('/profile', [ProfileController::class, 'destroy'])->name('profile.destroy');
 
-    // Admin routes (consolidated)
+    // Admin routes
     Route::prefix('admin')->middleware(['auth', 'admin'])->group(function () {
         Route::get('/', [AdminController::class, 'index'])->name('admin.index');
         Route::get('/create', [AdminController::class, 'create'])->name('admin.create');
         Route::post('/store', [AdminController::class, 'store'])->name('admin.store');
         Route::get('/edit/{book}', [AdminController::class, 'edit'])->name('admin.edit');
         Route::put('/update/{book}', [AdminController::class, 'update'])->name('admin.update');
-        Route::post('/borrow-status/{borrowedBook}', [AdminController::class, 'updateBorrowStatus'])->name('admin.updateBorrowStatus');
+        Route::put('/borrow-status/{borrowedBook}', [AdminController::class, 'updateBorrowStatus'])->name('admin.updateBorrowStatus');
         Route::post('/mark-as-paid/{borrowedBook}', [AdminController::class, 'markAsPaid'])->name('admin.markAsPaid');
         Route::get('/adjust-stock/{book}', [AdminController::class, 'adjustStock'])->name('admin.adjustStock');
         Route::post('/update-stock/{book}', [AdminController::class, 'updateStock'])->name('admin.updateStock');
-    
-        Route::get('/books', [BookController::class, 'adminIndex'])->name('admin.books.index');
-        Route::get('/books/create', [BookController::class, 'create'])->name('admin.books.create');
-        Route::post('/books', [BookController::class, 'store'])->name('admin.books.store');
-        Route::get('/books/{book}/edit', [BookController::class, 'edit'])->name('admin.books.edit');
-        Route::put('/books/{book}', [BookController::class, 'update'])->name('admin.books.update');
-        Route::patch('/books/{book}/toggle', [BookController::class, 'toggleStatus'])->name('admin.books.toggle');
-    
-        Route::get('/genres/create', [GenreController::class, 'create'])->name('admin.genres.create');
-        Route::post('/genres/store', [GenreController::class, 'store'])->name('admin.genres.store');
-        Route::get('/genres/{genre}/edit', [GenreController::class, 'edit'])->name('admin.genres.edit');
-        Route::put('/genres/{genre}', [GenreController::class, 'update'])->name('admin.genres.update');
-        Route::delete('/genres/{genre}', [GenreController::class, 'destroy'])->name('admin.genres.destroy');
+
+        // Genre routes
+        Route::get('/genres/create', [App\Http\Controllers\Admin\GenreController::class, 'create'])->name('admin.genres.create');
+        Route::post('/genres', [App\Http\Controllers\Admin\GenreController::class, 'store'])->name('admin.genres.store');
+        Route::get('/genres/{genre}/edit', [App\Http\Controllers\Admin\GenreController::class, 'edit'])->name('admin.genres.edit');
+        Route::put('/genres/{genre}', [App\Http\Controllers\Admin\GenreController::class, 'update'])->name('admin.genres.update');
+        Route::delete('/genres/{genre}', [App\Http\Controllers\Admin\GenreController::class, 'destroy'])->name('admin.genres.destroy');
     });
-    
+
+
     // Dashboard borrow route
     Route::post('/dashboard/borrow/{book}', [DashboardController::class, 'borrow'])->name('dashboard.borrow');
 });
